+++ Changelog.md
@@ -11,6 +11,9 @@ The next release will include the following feature enhancements and bug fixes:
 **Bug fixes**
 - Fixed search UI crashing on node select/preview (https://github.com/aws/graph-explorer/pull/177)
 
+**Bug fixes**
+- Fixed Gremlin/openCypher matching ID property on all keyword searches (https://github.com/aws/graph-explorer/pull/169)
+
 ## Release 1.3.1
 
 This patch release includes bugfixes for Release 1.3.0.